+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/authn/requests/UserAuthenticatorsDeleteRequest.java
@@ -9,21 +9,30 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class UserAuthenticatorsDeleteRequest extends BaseRequest {
 
+	/** An ID for an authenticator. */
 	@JsonProperty("authenticator_id")
 	private String authenticatorID;
 
+	/** The identity of a user or a service. */
 	@JsonInclude(Include.NON_NULL)
 	@JsonProperty("id")
 	private String id;
 
+	/** An email address. */
 	@JsonInclude(Include.NON_NULL)
 	@JsonProperty("email")
 	private String email;
+
+	/** A username. */
+	@JsonInclude(Include.NON_NULL)
+	@JsonProperty("username")
+	private String username;
 
 	private UserAuthenticatorsDeleteRequest(Builder builder) {
 		this.authenticatorID = builder.authenticatorID;
 		this.id = builder.id;
 		this.email = builder.email;
+		this.username = builder.username;
 	}
 
 	public static class Builder {
@@ -31,6 +40,7 @@ public class UserAuthenticatorsDeleteRequest extends BaseRequest {
 		private String authenticatorID;
 		private String id = null;
 		private String email = null;
+		private String username = null;
 
 		public Builder(String authenticatorID) {
 			this.authenticatorID = authenticatorID;
@@ -43,6 +53,16 @@ public class UserAuthenticatorsDeleteRequest extends BaseRequest {
 
 		public Builder setEmail(String email) {
 			this.email = email;
+			return this;
+		}
+
+		/**
+		 * Add a username to the request parameters.
+		 * @param username A username.
+		 * @return Builder.
+		 */
+		public Builder setUsername(String username) {
+			this.username = username;
 			return this;
 		}
 
@@ -61,5 +81,9 @@ public class UserAuthenticatorsDeleteRequest extends BaseRequest {
 
 	public String getEmail() {
 		return email;
+	}
+
+	public String getUsername() {
+		return username;
 	}
 }